BUNFOE: Implement init
This allows instantiating brand new instances that aren't unpacked from binary data, and are instead filled with the defaults for each field. Fields with no default specified will be defaulted to a sentinel value called UNREAD.

This also allows making copies of existing instances, and changing only some field values in the process.